Restructure player and users folders
