+++ front-end/ownFIFA/src/app/views/dashboard/players/players-import/players-import.component.html
@@ -1,0 +1,1 @@
+<p>players-import works!</p>
