+++ front-end/ownFIFA/src/app/views/dashboard/players/players-import/players-import.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-players-import',
+  standalone: true,
+  imports: [],
+  templateUrl: './players-import.component.html',
+  styleUrl: './players-import.component.scss'
+})
+export class PlayersImportComponent {
+
+}
