+++ front-end/ownFIFA/src/app/views/users/login/login/login.component.ts
@@ -5,7 +5,7 @@ import {
   ReactiveFormsModule,
   Validators,
 } from '@angular/forms';
-import { UserService } from '../../../services/user.service';
+import { UserService } from '../../../../services/user.service';
 import { Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
