+++ front-end/ownFIFA/src/app/views/users/login/logout/logout.component.scss
